pyproject.toml
[project]
name = "clash-royale-python"
version = "0.1.2"
description = "A synchronous, fully-typed Python library for the Clash Royale API"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Guillaume Coussot" },
]
keywords = ["clash-royale", "api", "wrapper", "supercell", "gaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
]

[project.urls]
Homepage = "https://github.com/guiepi/clash-royale-python"
Documentation = "https://guiepi.github.io/clash-royale-python/"
Repository = "https://github.com/guiepi/clash-royale-python"
Issues = "https://github.com/guiepi/clash-royale-python/issues"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=4.1.0",
    "ty>=0.0.1a0",
    "ruff>=0.1.0",
    "python-dotenv>=1.2.1",
    "pre-commit>=4.5.1",
    "pytest-recording>=0.13.4",
    "pytest-asyncio>=0.24.0",
]
docs = [
    "sphinx>=7.2.0",
    "furo>=2024.1.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "myst-parser>=2.0.0",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-design>=0.5.0",
    "sphinx-llms-txt>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/clash_royale"]
